Fix root-cause blockers from round 3 review: lossless arbitrary-precision exponents, bounded recursion, rawToken injection safety, native float encoding, nonfinite/underflow decode rejection, correct superDecoder(), enforced CreateGameRequest decode invariant, CI fixture provenance, and canonical-codec-only fixture tests
Addresses all 2 HIGH + 7 MEDIUM blockers from the round-3 independent
cumulative review of dcf1c08:

1. HIGH: extreme-exponent inputs no longer trap or allocate unboundedly.
   Introduced DecimalExponent, an arbitrary-range signed decimal exponent
   (digit-string arithmetic, no fixed-width Int assumptions) used
   throughout JSONNumber for arithmetic, canonicalization, hashing,
   Int conversion, and bounded scientific-notation serialization.

2. HIGH: unbounded nesting no longer stack-terminates. Added a single
   shared maxNestingDepth guard applied consistently to the parser,
   serializer, and JSONValue's own Codable conformance. Empirically
   rediscovered and lowered the limit from 512 to 64 after finding the
   recursive serializer overflows a constrained (Swift Testing worker
   thread) stack around depth ~400 -- far below the original limit's
   assumed safety margin on an 8MB main-thread stack.

3. MEDIUM: removed the public rawToken construction path entirely; only
   the parser-exclusive JSONNumber.parsed(...) can attach one, closing
   the JSON structural-injection surface.

4. MEDIUM: init(double:)/init(float:) no longer route through Decimal.
   They use Double/Float's own shortest round-tripping description,
   parsed through the safe JSONNumber grammar, preserving -0.0 and full
   magnitude range while explicitly rejecting NaN/infinity.

5. MEDIUM: decoding to Double/Float now rejects nonfinite results and
   silent underflow-to-zero for a nonzero source coefficient, while
   still accepting genuine (correctly signed) zero at any exponent.

6. MEDIUM: fixed both bare and keyed superDecoder()/superEncoder() to
   read/write under the conventional "super" key (and correct keyed
   index), matching standard Codable super-container semantics, plus a
   related nestedContainer/nestedUnkeyedContainer codingPath-index bug.

7. MEDIUM: CreateGameRequest's decoder now enforces the same
   campaign/scenario sum invariant the encoder already enforced, via a
   new closed CampaignOrScenario type shared by both directions, so an
   invalid combination can no longer be represented at all.

8. MEDIUM: added Scripts/verify-contract-fixture-provenance.sh plus a
   mise task and CI job that fetch the pinned backend commit
   6a1befbd7b01b4a0f763e41260ae4dd1a5d14c27 from djensenius/ArkhamHorror
   and byte-compare every governed contract fixture against it, so
   fixture drift is caught independent of locally-recomputable hashes.

9. MEDIUM: migrated remaining contract-fixture test suites off stock
   JSONDecoder/JSONEncoder onto the canonical ContractJSON path, and
   added aggregate-fixture lossless-numeric-passthrough tests (long
   integer id, 1e128) through a real request type's decode/encode
   cycle, not just isolated JSONNumber unit tests.

Adversarial test coverage added for every item above (extreme exponent
triggers at/beyond Int64 boundaries, nesting-depth boundary tests,
rawToken injection/self-consistency tests, Double/Float boundary
round-trips including subnormals and -0.0, nonfinite/underflow decode
rejection, a 3-level class hierarchy superDecoder/superEncoder
round-trip with coding-path and write-back-error-propagation
assertions, and full presence/null/absent decode-time coverage of the
campaign/scenario invariant). New test files were split out where
needed to stay within SwiftLint's file/type-length limits.

Verified: swiftformat (0/171 files need formatting), swiftlint --strict
(0 violations), swift test (661 tests / 43 suites passed), xcodegen
project-check (no drift), contract-fixture-provenance script (all
fixtures match backend commit 6a1befbd7b01b4a0f763e41260ae4dd1a5d14c27),
and all 4 platform builds (iOS/macOS/tvOS/visionOS) succeeded, all with
DEVELOPER_DIR=/Applications/Xcode-beta.app/Contents/Developer.

+++ Packages/ArkhamHorrorShared/Sources/ArkhamHorrorShared/Domain/Contract/LosslessJSONEncoding.swift
@@ -103,7 +103,7 @@ struct LosslessJSONKeyedEncodingContainer<Key: CodingKey>: KeyedEncodingContaine
     }
 
     mutating func encode(_ value: Float, forKey key: Key) throws {
-        try set(.leaf(.number(JSONNumber(double: Double(value)))), forKey: key)
+        try set(.leaf(.number(JSONNumber(float: value))), forKey: key)
     }
 
     mutating func encode(_ value: Int, forKey key: Key) throws {
